docs(solar_system_stage_2):README markdown 기능 목록, 풀이 과정, 구현 결과 업데이트

+++ SolarCircle.java
@@ -59,7 +59,7 @@ public class SolarCircle implements Circle {
     private double calculateDistance(int x, int y, double radius) {
         double distance = Math.abs(
                 radius - Math.sqrt(
-                        (Math.pow(x - xLocation, 2) + Math.pow(y - yLocation, 2))));
+                        (Math.pow(x - (int)xLocation, 2) + Math.pow(y - (int)yLocation, 2))));
 
         return distance;
     }
